client/requirements.txt
grpcio-tools
protobuf3-to-dict
Flask
werkzeug
functions-framework==3.*
twilio